now editing profile messages is working, deleting profile messages has been deleted

+++ utils/clear_back.py
@@ -19,6 +19,8 @@ async def clear_back(bot, message, start_num=0, anchor_message=None):
                 continue
         if message.message_id - i in temp_storage.exceptions_messages:
             continue
+        if message.message_id - i == temp_storage.profile_message:
+            continue
         asyncio.create_task(
             delete_wrap_by_call(bot.delete_message(message.chat.id, message.message_id - i))
         )
